refactor(notebook): import shared state stores directly
