+++ apps/notebook/src/components/MarkdownCell.tsx
@@ -31,7 +31,7 @@ import {
   useIsNextCellFromFocused,
   useIsPreviousCellFromFocused,
   useSearchQuery,
-} from "../lib/cell-ui-state";
+} from "@/components/notebook/state/cell-ui-state";
 import { onEditorRegistered, onEditorUnregistered } from "../lib/cursor-registry";
 import { registerCellEditor, unregisterCellEditor } from "../lib/editor-registry";
 import { logNotebookIsolatedDiagnostic } from "../lib/isolated-diagnostics";
