+++ apps/notebook/src/components/NotebookView.tsx
@@ -31,14 +31,14 @@ import { cn } from "@/lib/utils";
 import type { TracebackCellTarget } from "@/components/outputs/traceback-output";
 import { usePresenceContext } from "../contexts/PresenceContext";
 import { EditorRegistryProvider, useEditorRegistry } from "../hooks/useEditorRegistry";
-import { useFocusedCellId, useSearchCurrentMatch } from "../lib/cell-ui-state";
+import { useFocusedCellId, useSearchCurrentMatch } from "@/components/notebook/state/cell-ui-state";
 import { logger } from "../lib/logger";
 import {
   getCellById,
   getNotebookCellsSnapshot,
   useCell,
   useMaterializeVersion,
-} from "../lib/notebook-cells";
+} from "@/components/notebook/state/cell-store";
 import {
   getCellOutputsSnapshot,
   subscribeOutputsVersion,
